requirements.txt
Flask
requests
pandas
SQLAlchemy
psycopg2-binary
python-dotenv
